+++ midealocal/devices/cc/__init__.py
@@ -173,7 +173,7 @@ class MideaCCDevice(MideaDevice):
 
     def set_target_temperature(
         self,
-        target_temperature: int,
+        target_temperature: float,
         mode: int | None,
         zone: int | None = None,  # noqa: ARG002
     ) -> None:
